Modularized code using cogs
